+++ CHANGELOG.md
@@ -93,7 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Used by default in Neovim >= 0.10.
 - LSP: `:RustLsp testables` command, which is equivalent
   to `:RustLsp runnables`, but filters the runnables for tests only,
-  
+
 > [!IMPORTANT]
 >
 > In Neovim < 0.10, `'background'` executor blocks the UI while running tests.
@@ -107,7 +107,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and achieve an experience similar to Rust Playground.
   (currently requires a nightly compiler).
   Thanks [saying121](https://github.com/saying121)!
-- Config: `tools.rustc_unpretty` arguments for `rustc`.
+- Config: `tools.rustc` arguments for `rustc`.
 
 ### Changed